Analyst Fix

+++ src/main/java/Business/Enterprise/DeliveryDepartment/Delivery.java
@@ -16,8 +16,8 @@ public class Delivery {
     CoffeeOrderRequest order;
     Destination destination;
     
-    int dateTimeSent;
-    int dateTimeArrived;
+    long dateTimeSent;
+    long dateTimeArrived;
     
     public Delivery(CoffeeOrderRequest o, Rider r) {
         order = o;
@@ -45,19 +45,19 @@ public class Delivery {
         return destination;
     }
 
-    public int getDateTimeSent() {
+    public long getDateTimeSent() {
         return dateTimeSent;
     }
 
-    public void setDateTimeSent(int dateTimeSent) {
+    public void setDateTimeSent(long dateTimeSent) {
         this.dateTimeSent = dateTimeSent;
     }
 
-    public int getDateTimeArrived() {
+    public long getDateTimeArrived() {
         return dateTimeArrived;
     }
 
-    public void setDateTimeArrived(int dateTimeArrived) {
+    public void setDateTimeArrived(long dateTimeArrived) {
         this.dateTimeArrived = dateTimeArrived;
     }
 }
